Solves error on edge-case
In some datasets of very fine sediment, all cases are classified as "mud". Because of the absence of Gravel component, the function `.texture_sedim()` has a fatal error because it can't find both `Gravel` and `sandmud` variables. This is solved by checking for their presence, and adding a dummy variable when required. This also solves the opposite edge case where a sample could be entirely Gravel.